move c++ framereader to selfdrive/ui/replay (#21053)
