--- tools/clib/.gitignore
@@ -1,1 +1,0 @@
-cframereader.cpp
